refactor(blog): 모듈 경계 위반 해소로 CI Bean 충돌/ArchUnit 실패 수정
## 왜
PR #14 CI에서 17개 테스트 연쇄 실패 + ArchitectureTest 실패.
원인 2가지가 얽혀 있었음:

1. ArchitectureTest 위반
   - 규칙: `noClasses().that().resideInAPackage("..common..")`
     .should().dependOnClassesThat().resideInAnyPackage("..blog..", ...)
   - `..common..` 와일드카드가 `com.study.blog.common`도 매칭
   - 블로그 팀이 새로 만든 `com.study.blog.common.*`가 같은 blog 하위 참조 시
     규칙 위반으로 판정됨

2. ConflictingBeanDefinitionException (17 AuthIntegrationTest 연쇄 사망)
   - `BlogTestApplication` @SpringBootApplication이 기본 scanBasePackages로
     부팅 → 클래스패스의 다른 모듈 빈까지 스캔 시도
   - app 모듈 테스트와 Bean 이름 중복 등록으로 컨텍스트 로드 실패

## 무엇을
1. `com.study.blog.common` → `com.study.blog.shared` 패키지 리네임
   - 5개 파일 이동: ApiResponse, MockAuth, BlogErrorCode, BlogException,
     GlobalExceptionHandler
   - 참조한 10개 파일의 import도 일괄 수정

2. `BlogTestApplication` 수정 — 삭제 대신 스캔 범위 제한
   - `@SpringBootApplication(scanBasePackages = "com.study.blog")` 명시
   - 기본값으로 두면 다른 모듈까지 스캔해서 위 Bean 충돌 재현됨
   - 주의 주석으로 이유를 박아둠 (재발 방지)

## 검증
- ./gradlew compileJava 성공
- ./gradlew :app:test 성공 (기존 17개 실패 전부 복구)
- ArchitectureTest 통과
