+++ blog/src/main/java/com/study/blog/admin/AdminController.java
@@ -3,9 +3,9 @@ package com.study.blog.admin;
 import com.study.blog.admin.dto.AdminPostResponse;
 import com.study.blog.admin.dto.AdminStatsResponse;
 import com.study.blog.admin.dto.PostStatusUpdateRequest;
-import com.study.blog.common.auth.MockAuth;
-import com.study.blog.common.exception.BlogErrorCode;
-import com.study.blog.common.exception.BlogException;
+import com.study.blog.shared.auth.MockAuth;
+import com.study.blog.shared.exception.BlogErrorCode;
+import com.study.blog.shared.exception.BlogException;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
