+++ blog/src/main/java/com/study/blog/admin/AdminService.java
@@ -3,8 +3,8 @@ package com.study.blog.admin;
 import com.study.blog.admin.dto.AdminPostResponse;
 import com.study.blog.admin.dto.AdminStatsResponse;
 import com.study.blog.comment.CommentRepository;
-import com.study.blog.common.exception.BlogErrorCode;
-import com.study.blog.common.exception.BlogException;
+import com.study.blog.shared.exception.BlogErrorCode;
+import com.study.blog.shared.exception.BlogException;
 import com.study.blog.post.PostLikeRepository;
 import com.study.blog.post.PostRepository;
 import com.study.blog.post.PostTagRepository;
