+++ blog/src/main/java/com/study/blog/comment/CommentService.java
@@ -3,8 +3,8 @@ package com.study.blog.comment;
 import com.study.blog.comment.dto.CommentCreateRequest;
 import com.study.blog.comment.dto.CommentResponse;
 import com.study.blog.comment.dto.CommentUpdateRequest;
-import com.study.blog.common.exception.BlogErrorCode;
-import com.study.blog.common.exception.BlogException;
+import com.study.blog.shared.exception.BlogErrorCode;
+import com.study.blog.shared.exception.BlogException;
 import com.study.common.entity.Comment;
 import com.study.common.entity.CommentLike;
 import java.util.ArrayList;
