+++ blog/src/main/java/com/study/blog/shared/ApiResponse.java
@@ -1,4 +1,4 @@
-package com.study.blog.common;
+package com.study.blog.shared;
 
 public class ApiResponse<T> {
