+++ blog/src/main/java/com/study/blog/shared/exception/BlogErrorCode.java
@@ -1,4 +1,4 @@
-package com.study.blog.common.exception;
+package com.study.blog.shared.exception;
 
 import org.springframework.http.HttpStatus;
